Show 'madatory asterisk' for all fields when applicable
Restructuring the edit template a bit, because $mandatory will only be
applied in the format section.
Also the sweetalert will only pick up the title, if it is in the same
parent as the input triggering the message, so all the extra spans have
to go away.

Changed color of the asterisk for the title, because red on blue does
not look nice.

Note: Not changing %MAKETEXT{"Due date"}% to $xlatedescription for
compatibility reasons.

Ticket: #15720
Release-Note(fix, appdev) {
  Will now correctly mark mandatory fields in the task panel.
}

Release-Note(fix, appdev) {
  The alert for missing mandatory fields will now correctly pick up the
  title of the field.
}

Release-Note(change, appdev) {
  Changed the colour of the asterisk indicating the title field as
  mandatory to white.
[...]
  The default red one clashes with the blue background.
}

Release-Note(fix, admin) {
  The asterisk indicating the title field as mandatory will no longer
  move to the next line when creating a new task.
[...]
  Some apps may need to be updated.
}